Add licenses and copyright info
While the legal questions of Flathub are being worked out this takes
some basic steps to ensure license compliance.

Most of the time, license compliance is as simple as including the
license and/or copyright file with distribution of the application.
In traditional linux packaging it is common to have a post install
step that does this. Fortunately flatpak-builder can do this too.

While this works it is not ideal. It does add a lot more manual work
and it does not help in keeping the manifest minimal.

Some ideas to improve the current workflow:
- Because most legal files follow some sort of naming convention
  flatpak-builder could include basic support for installing legal
  files automatically.
- The post-install step invalidates caches causing unnecessary
  rebuilds. It really should treated as its own step.