Update sdk to 3.26 and PulseEffects v3.0.1
The 3.26 runtime contains a newer version of gstreamer, compatible with
PulseEffects. This is now used instead.

However, some gstreamer and ladspa plugins still need to be built:

- libgstladspa.so (ladspa/gstreamer plugin bridge)
- fast_lookahead_limiter_1913.so
- sc4_1882.so

Instead of building everything to provide this, only the required
plugins are built. This greatly reduces build time and saves some space.

Unfortunately because gst-plugins-bad lack a global disable flag and the
ladspa plugin lacks meson support, this currently adds a lot of
config-opts to the manifest.

A patch that adds the required meson build support is in progress.

On a minor, unrelated note, the app-id is now used for the license
directory name, in case this is ever exported.