Update pulseeffects to v3.1.7
- Also update numpy to v1.14.0

The new plugins required by pulseeffects are:

- ladspa-ladspa-rubberband-so-rubberband-pitchshifter-stereo
  (Pitch Shifting from Rubber Band Library)
- lsp-plug-in-plugins-lv2-comp-delay-x2-stereo
  (Delay Compensation from Linux Studio Plugins Project)

Unfortunately the second plugin cannot be included at this time.

Due to the financing model of the project, only some of the plugins are
open source. In itself that is a non-issue, since the one plugin
required at this time is open source.

The bigger issue is the source code only being available as a svn
repository, which `flatpak-builder` does not support.