gfortran: Use org.freedesktop.Sdk.Extension.gcc7
Builds of the "org.freedesktop.Sdk.Extension.gfortran-62" extension for
arm and aarch64 remain unavailable. This is mainly due to issues with
said architectures and GCC 6. It seems this is unlikely to change.

GCC 7 is now available as an extension. This extension include gfortran
support, and most importantly it offers builds for arm and aarch64.