gfortran: Install libgcc from sdk extension
Building with gcc 7 but linking against version 6.x of libgcc_s.so is a
problem. At least with i386, that requires a function `__divmoddi4`,
that is otherwise unavailable.

Ignoring this issue breaks numpy, and modules that depend on it, like
scipy. Error while importing numpy:

multiarray.cpython-35m-i386-linux-gnu.so: undefined symbol: __divmoddi4