Switch to yml file extension
This only changes the file extension from "json" to "yml" (YAML). It
does not include a functional YAML file.

The purpose of this commit is to ensure the git history of the manifest
file is kept. If the old manifest is removed the history is as well,
which apparently causes the flathub build bot to fail.

There may very well be better ways of doing this, but at least this
solves the bot issue.