fftw: Remove leftover fortran options
Flatpak makes all installed sdk extensions available during build.
This means it is possible to build, for example, fortran code without
specifying `org.freedesktop.Sdk.Extension.gfortran` as a sdk extension.

That is as long as the extension is already installed. On a clean host
system, where it is not, all attempts to build the flatpak will instead
fail.

So, lesson is: Don't forget things like this.