rubberband: Fix Makefile
The newer version includes some good and bad changes to its Makefile.
There is now a ladspa target, which simplifies building only what is
needed by pulseeffects. Unfortunately the install target does not work
well with this.

There is also an issue with missing directories that causes linking to
fail. That is easily patched.