calf: Fix failure to detect lv2 >= 1.16.0
Calf used to check for lv2core, but the latest version of lv2 broke this
by removing the .pc file providing this. As a result the calf lv2
plugins would not build and install, which causes PulseEffects to report
missing plugins.

Upstream git (master) has fixed this already. Fixed by backporting
upstream patch.