Remove RealtimeKit access
This never worked since processes inside the sandbox run in their own
pid namespace and RealtimeKit, which runs on the host, expects a pid in
the host namespace.

The best solution is likely to wait for a RealtimKit portal interface to
be developed. It would likely function similar to Feral Interactive's
GameMode portal.