Fix mda.lv2 build
The GNOME runtime, version 3.34, no longer include python2 since it will
soon be EOL and the last release of mda.lv2 uses a waf binary that does
not work python 3.7.

Since the release of mda.lv2 1.2.2 the master branch include little but
updates and fixes to the build system the simplest solution is to just
use the latest commit from master.